+++ CHANGELOG.md
@@ -4,9 +4,19 @@ for changelogs and semantic versioning available at:
 * https://keepachangelog.com/en/0.3.0/
 * https://semver.org/
 
+## [0.1.2] - 2018-06-27, 2018-07-05
+### Added
+- GraphQL, GraphiQL
+- seeds.rb
+- Character model, spec, factory
+- DatabaseCleaner, Faker, FactoryBot, RSpec
+
+### Changed
+- Rename characters.name to characters.title
+
 ## [0.1.0] - 2018-05-20
 ### Added
 - Initial project setup: git, documentation
 - rbenv/ruby 2.5.1
 - gems: bundler, rails 5.2.0
-- *postgresd
+- postgresql